[refactor] move function to specific doc_handler file
